Fix unresolved promise bestuursorgaanintijdselect
There is a bug in production that we are not able to reproduce
concerning the bestuursorgaan-in-tijd-select component. It has to be
refreshed to have all the options for Bestuursperiode. We think it might
come from an unresolved promise.
This is why here we change a bit the way we call the component by only
passing the full model to it and then await the properties we need.