Fix an issue with the inflector config registration
EmberData deprecated ember-inflector support in v5.3.4, but the implemenation had some timing issues with our config registration. Because we were using a side-effecting import the custom rules would  be registered before EmberData's compat detection code is run, which results in EmberData not using the ember-inflector based setup anymore.

By converting the side-effecting import into a function we call _after_ the initializers are run, we bypass the problem and the app works as expected (with deprecations being shown).